Сделана полная реализация комментариев+немного улучшена логика лайков
обязательно надо переделать flyway+почистить классы от лишних импортов

+++ src/main/java/dev/trendio_back/service/InitializerService.java
@@ -40,8 +40,8 @@ public class InitializerService {
                 .username(user.getUsername())
                 .address("hui")
                 .tags(tags)
-                .latitude(BigDecimal.valueOf(51.662856))
-                .longitude(BigDecimal.valueOf(39.1979146))
+                .latitude(51.662856)
+                .longitude(39.1979146)
                 .build());
     }
 }
